move some function into bnb.py
