+++ lycoris/utils/bnb.py
@@ -1,3 +1,5 @@
+from functools import cache
+
 try:
     from bitsandbytes.nn import (
         LinearNF4, Linear8bitLt, LinearFP4
@@ -10,3 +12,19 @@ except:
         pass
     class LinearFP4(nn.Linear):
         pass
+
+from ..logging import logger
+
+
+QuantLinears = (
+    Linear8bitLt,
+    LinearFP4,
+    LinearNF4
+)
+
+
+@cache
+def log_bypass():
+    return logger.warning(
+        "Using bnb with LyCORIS will enable force-bypass mode."
+    )
